# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking AI - API Documentation",
        "description": "API REST para el sistema de gestión de parqueaderos multi-sucursal. Permite administrar sucursales, usuarios, vehículos, tarifas, descuentos, impuestos y operaciones de entrada/salida de vehículos.",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@parking-ai.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "http://localhost:8001/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Listar usuarios",
                "operationId": "2979f0297ab881ad0b8b86ee613f077b",
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear usuario",
                "description": "Crear usuario",
                "operationId": "7601d6cfa9768b3857d0e58d45eb618b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "document_number",
                                    "document_type_id",
                                    "role_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 80,
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "password123"
                                    },
                                    "document_number": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "1234567890"
                                    },
                                    "document_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "3001234567"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado exitosamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener usuario",
                "description": "Obtener usuario por ID",
                "operationId": "bc1bd7b0a9c5c03d14165e230c9b74a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrado exitosamente"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "description": "Actualizar usuario",
                "operationId": "a13246ab4f01cf3913b77f7d4520bbd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "last_name",
                                    "email",
                                    "document_number",
                                    "document_type_id",
                                    "role_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 80
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 80
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "document_number": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "document_type_id": {
                                        "type": "integer"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "role_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado exitosamente"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario",
                "description": "Eliminar usuario",
                "operationId": "edba4ea5d13c283794da36b1eb31b2fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado exitosamente"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Laravel Sanctum. Obtén el token usando el endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}